Remove ImageType
